LyricSystem a1.49.01
Preparing to bridge to a new style of system, this one will use hosted files with lyrics arranged in folders (by artist) which will be fetched and displayed in one div.
This should be a superior alternative to having a thousand+ -line .html file. It becomes too cumbersome to add new songs to that.
Also added in the Iuliia transcription service, which works flawlessly :) thanks guys!.

You can see these alpha changes in new.html before I make the overhaul.

a1.5 will be my confimation update of the new system's operation, and it will continue to progress from there with CSS changes until 2.0 drops (this will be the migration of all current + planned lyrics into this new system, on (maybe) one HTML page).