LyricSystem a1.49.02
Fix 1 of the general CSS for the new system (new.html). Pushing so I can phone-test tomorrow.

Created lyricsButtonAssembler.py: In the future, this will go through the /txt/ folder and generate the HTML required for the song list for the slavic folder.  In the future, this will be for all folders.

Created lyricsTransformer.py: In the future, this will be able to run through all .txt files and transform them into more HTML-friendly versions. In the past, I had to use <pre>s in order to get the lyric formatting proper-like and I don't like that. We're doing this smarter from now on, and we're automating that work with Python.

Updated pages to represent the 1.49 version of the site.

Moved the three folders of slavic music to the /txt/slavic/ folder.
My plan is to have a lyrics page for Slavic music, for Vocaloid/Japanese music, and for English/other music. So, there will be a main folder in /txt/ for each of them.