CSS Fixes and New Songs | LyricSystem a1.49.03
Fix 2 of the CSS for new.html. I am very quaint with the results. I'm no UX guy and I'm just making this for myself, so my perfectionism is limited. Hoorah for now!!

Compiled all of the Tima lyrics into respective .txt's. Hooray!

Made the "Status" empty link navigate to new.html. Changed name to "New System".

Next update will be the bridge update to see how the system works in-field (meaning when I'm out and about, trying to find lyrics): I'll actually add the columns/buttons for all of the Tima and Rauf+Faik songs.
I'm just gonna finish that python script to build the list for me. Easy stuff in concept :)

I also want to build a script that will just build this website for me. So, instead of tweaking one page and having to update it across all of them, it's basically just going to do what lyricsButtonAssembler.py will do, but on a mass scale with extra steps: For each section of the /txt/ folder, It will generate a boilerplate LyricSystem page for the folder.
This makes my job easier for when I find songs I like in the future. Then, I can just pop the lyrics in a .txt, run this file, and push the commit!