Slavic System Finalization New Song | LyricSystem a1.49.07
I have turned the new system into slavic.html, which is now linked everywhere on the site instead of new.html (which has been months deprecated iirc).
In relation to this, I updated makeSlavic.py (and updated lyricsButtonAssembler.py to match) so that it automatically walks through all the Slavic lyrics and builds the .html file for me (instead of having to copy-paste into the file, which is stupid).
This update allows me to very quickly add new songs as a .txt, double-click that python file, and push to GitHub.

:)