Slavic Songs System Release | LyricSystem a1.5.0
The last update was erroneously titled a1.49.07, when it was actually a1.49.09. Either way, we have entered a1.5.0 now!! Woo!

This update constitutes the full release of my new Slavic Songs system, which includes (as of now) five artists neatly arranged in one page with lyrics of 65 songs. Also included is the ability to resize the lyrics for your screen, and to transliterate lyrics from Cyrillic alphabet to roman characters.
In the background, we have multiple .py files which work together to make the automated updating of this simple HTML page very easy for me! All I have to do is add new lyrics, and run makeSlavic.py.

In the future, this system will be expanded to the rest of my site, with one Python file rebuilding all of the pages from one single template.